requirements.txt
Flask
tensorflow
numpy
Pillow
streamlit
requests
gunicorn
